Only handle form submit event when form not valid
A form may have other event handlers attached to the `submit` event.
For example, when implementing an Ajax form. Rather than call the
`submit` method directly which bypasses those handlers, if the form is
valid, this allows the original form submit event to propagate.

However, if the form is invalid, this stops propagation of the form
event entirely, even to other registered event handlers.

This requires that client validation registers its event handler on
forms first.