Check for special checkbox hidden input
Browsers do not send a value for unchecked checkboxes. To work around this, platforms such as ASP.NET render a hidden input with the same name as the checkbox so that a value ("false") is still submitted even when the checkbox is not checked. We check this special case here.

Fixes #48